+++ Sources/KeyboardKit/Gestures/KeyboardGestures.swift
@@ -64,7 +64,7 @@ struct KeyboardGestures<Content: View>: View {
             Color.clearInteractable
                 .gesture(dragGesture(for: geo))
                 .simultaneousGesture(tapGesture)
-                .simultaneousGesture(doubleTapGesture)
+                .withOptionalGesture(doubleTapGesture)
                 .simultaneousGesture(longPressGesture)
                 .simultaneousGesture(longPressDragGesture(for: geo))
         })
@@ -79,10 +79,12 @@ private extension KeyboardGestures {
     /**
      This is a plain double-tap gesure.
      */
-    var doubleTapGesture: some Gesture {
-        TapGesture(count: 2)
-            .onEnded { doubleTapAction?() }
+    var doubleTapGesture: _EndedGesture<TapGesture>? {
+        guard let action = doubleTapAction else { return nil }
+        return TapGesture(count: 2)
+            .onEnded { action() }
     }
+
     
     /**
      This is a drag gesture that starts immediately.
@@ -189,5 +191,20 @@ private extension KeyboardGestures {
     
     func stopRepeatTimer() {
         repeatTimer.stop()
+    }
+}
+
+
+// MARK: - Provate View Extension
+
+private extension View {
+    
+    @ViewBuilder
+    func withOptionalGesture<GestureType: Gesture>(_ gesture: GestureType?) -> some View {
+        if let gesture = gesture {
+            self.gesture(gesture)
+        } else {
+            self
+        }
     }
 }